Credibility/Enemy AI update
I've made a draft of the credibility system where each time the player calls (checks win condition) they lose credibility the amount they lose scales on how many calls they've made. The loss of credibility can be lessened if the player has a high amount of points. Also the enemy will now switch to the flee state when seen by the player and can damage the player if it attacks from behind.